Add regression tests for the expired default chain and the multidimensional reconciliation
`t_nested_expired_default_chain` forces packed part storage, hence a horizontal
merge, so the merge has to fill the defaults of two expired columns one of which
reads the other. `t_nested_filtered_default_2d` covers reconciling a constant
default of an `Array(Array(String))` subcolumn against the shared `Nested`
offsets, where only the outermost level is shared.

+++ src/Storages/MergeTree/IMergeTreeReader.cpp
@@ -442,7 +442,7 @@ void IMergeTreeReader::filterSharedOffsetsOfMissingDefaults(const ColumnPtr & fi
         const auto & innermost_offsets = assert_cast<const ColumnUInt64 &>(*offsets.back()).getData();
         size_t data_size = innermost_offsets.empty() ? 0 : innermost_offsets.back();
 
-        ColumnPtr shape = ColumnUInt8::create(data_size, 0);
+        ColumnPtr shape = ColumnUInt8::create(data_size, UInt8(0));
         for (auto it = offsets.rbegin(); it != offsets.rend(); ++it)
             shape = ColumnArray::create(shape, *it);
